libusb: implement `libusb_get_parent`
Summary: Newer versions of drivers such as libwacom (graphics tablets) or libfprint (fingerprint scanners) call g_usb_device_get_parent. This in turn calls libusb_get_parent on platforms which implement it, and returns NULL on platforms that don't. This patch implements this function on FreeBSD.

Test Plan: Tried different combinations of different devices connected to different hubs and it all seems to work. Tried on the AMD Framework 13. Used an ad-hoc program to test this but I'd eventually like to add a tree view kind of thing to usbconfig like in lsusb.

Reviewers: danfe

Subscribers: imp

Differential Revision: https://reviews.freebsd.org/D46992