USB4 acknowledge hotplug events
Summary: When getting a hotplug event from the hardware, transmit a hotplug acknowledgment packet. This is needed even if we don't do anything with the hotplug events, as the hardware will continuously "spam" us with hotplug events otherwise.

Test Plan: Tested on AMD Pink Sardine controller (1022:1668/9), which is host interface version 1.0.

Reviewers: mohammad_fbsdphab_thelightbird.com, imp, jhb, emaste

Subscribers: imp

Tags: #usb

Differential Revision: https://reviews.freebsd.org/D49454