USB4 Pink Sardine controller support
Summary:
Add support for the AMD Pink Sardine controller (1022:1668/9) and fix issue where we weren't getting interrupts for more than the first command sent on the ring (which presumably was not an issue for the hardware Scott was testing on).

I also added a couple comments as notes-to-self which I'll come back to.

Test Plan: Testing on AMD Pink Sardine controller (1022:1668/9), which is host interface version 1.0.

Reviewers: mohammad_fbsdphab_thelightbird.com, imp, jhb, emaste

Subscribers: imp

Tags: #usb

Differential Revision: https://reviews.freebsd.org/D49451