USB4 initial work on suspend routine
Summary:
Initial work on a suspend routine for USB4 on v1.0 host interfaces. At the moment I'm not quite concerned about resume, as it would already be quite nice to be able to suspend USB4 routers as this is a prerequisite for S0i3 on AMD.

This will also work on v2.0, but we might want to handle this differently in the future for these hosts.

Test Plan: Tested on AMD Pink Sardine controller (1022:1668/9), which is host interface version 1.0. The SLPR (sleep ready) bit is being set by the hardware.

Reviewers: mohammad_fbsdphab_thelightbird.com, imp, jhb, emaste

Subscribers: imp

Tags: #usb

Differential Revision: https://reviews.freebsd.org/D49453