acpi: Add back `hw.acpi.suspend_state` sysctl
When writing an ACPI S-state to it it will set kern.power.suspend to the
appropriate sleep type, and when reading from it it will return the
corresponding ACPI S-state to the sleep type in kern.power.suspend.

This is deprecated and kern.power.suspend should be used directly
instead, but add this back because zzz(1) makes use of this and we can't
easily rewrite it just now.

PR:		289634
Sponsored by:	The FreeBSD Foundation

Reviewed by:	markj
Differential Revision:	https://reviews.freebsd.org/D52600